Move SpellTimer and SpellTimerJSON to new file
